[Feat] 방 생성, 조회 api 구현

See merge request s13-webmobile1-sub1/S13P11A404!21

+++ frontend/BE/src/main/java/com/a404/duckonback/dto/UserDTO.java
@@ -18,5 +18,5 @@ public class UserDTO {
     private String role;
     private String language;
     private String imgUrl;
-    private List<Integer> artistList;
+    private List<Long> artistList;
 }
